requirements.txt
# Data Manipulation & Extraction
pandas>=2.0.0
requests>=2.31.0

# Google Cloud Platform Client Libraries
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0

# Dependencies for BigQuery and Pandas performance
db-dtypes>=1.2.0
pyarrow>=12.0.0

# Optional: dbt adapter for local development (not required if using dbt Cloud)
# dbt-bigquery>=1.6.0